add sell management

+++ eStoreManager-SERVER/src/main/java/com/example/store/payload/product_management/request/UpdateProductRequest.java
@@ -18,7 +18,7 @@ public class UpdateProductRequest {
     private String barcode;
 
     @Min(0)
-    private Long quantities;
+    private Float quantities;
 
     public String getName() {
         return name;
@@ -36,7 +36,7 @@ public class UpdateProductRequest {
         return barcode;
     }
 
-    public Long getQuantities() {
+    public Float getQuantities() {
         return quantities;
     }
